+++ components/Settings.tsx
@@ -1,26 +1,6 @@
 import { use } from 'react';
 import { config } from '../lib/config';
-
-interface SettingsProps {
-  useTTS: boolean;
-  useInternet: boolean;
-  usePhotos: boolean;
-  useChainMode: boolean;
-  useJsonMode: boolean;
-  useAgentMode: boolean;
-  onTTSToggle: () => void;
-  onInternetToggle: () => void;
-  onPhotosToggle: () => void;
-  onChainModeToggle: () => void;
-  onJsonModeToggle: () => void;
-  onAgentModeToggle: () => void;
-  setTTS: (useTTS: boolean) => void;
-  setInternet: (useInternet: boolean) => void;
-  setPhotos: (usePhotos: boolean) => void;
-  setChainMode: (useChainMode: boolean) => void;
-  setJsonMode: (useJsonMode: boolean) => void;
-  setAgentMode: (useAgentMode: boolean) => void;
-}
+import { SettingsProps } from '../types/components';
 
 export const Settings: React.FC<SettingsProps> = ({
   useTTS,
