+++ types/responses.ts
@@ -1,0 +1,16 @@
+export interface Responses {
+  fruitsVegetables: string;
+  fastFood: string;
+  waterIntake: string;
+  sugaryBeverages: string;
+  alcohol: string;
+  exerciseDays: string;
+  exerciseDuration: string;
+  sleepHours: string;
+  stressFrequency: string;
+}
+
+export interface Response {
+  question: string;
+  answer: string;
+}
